fix setup.py

+++ setup.py
@@ -28,28 +28,29 @@ if travis:
     with open('django_rediser/config.py', 'w', encoding="utf-8") as f:
         f.write("__version__ = '{}'".format(version))
 
-try:
-    import pypandoc
-
-    print("Converting README...")
-    long_description = pypandoc.convert('README.md', 'rst')
-    if branch:
-        long_description = long_description.replace('django-rediser.svg?branch=master', 'django-rediser.svg?branch={}'.format(branch))
-
-except (IOError, ImportError, OSError):
-    print("Pandoc not found. Long_description conversion failure.")
-    with open('README.md', encoding="utf-8") as f:
-        long_description = f.read()
-else:
-    print("Saving README.rst...")
+if os.path.isfile('README.md'):
     try:
-        if len(long_description) > 0:
-            with open('README.rst', 'w', encoding="utf-8") as f:
-                f.write(long_description)
-            if travis:
-                os.remove('README.md')
-    except:
-        print("  failed!")
+        import pypandoc
+
+        print("Converting README...")
+        long_description = pypandoc.convert('README.md', 'rst')
+        if branch:
+            long_description = long_description.replace('django-rediser.svg?branch=master', 'django-rediser.svg?branch={}'.format(branch))
+
+    except (IOError, ImportError, OSError):
+        print("Pandoc not found. Long_description conversion failure.")
+        with open('README.md', encoding="utf-8") as f:
+            long_description = f.read()
+    else:
+        print("Saving README.rst...")
+        try:
+            if len(long_description) > 0:
+                with open('README.rst', 'w', encoding="utf-8") as f:
+                    f.write(long_description)
+                if travis:
+                    os.remove('README.md')
+        except:
+            print("  failed!")
 
 setup(
     name='django-rediser',
